dependency injection working in unit test - production implementation WIP
