--- src/UserScores.js
@@ -1,12 +1,0 @@
-// hardcoded for now but will be an API call to our database, once built
-// sort in order of descending score for Scoreboard.jsx to render accurate leader order
-export const userScores = [
-  {
-    username: "octokitty", 
-    score: 100
-  },
-  {
-    username: "buck-git list", 
-    score: 50 
-  }
-];
