+++ src/components/ScoreBoard/ScoreBoard.jsx
@@ -1,19 +1,6 @@
-import React, { useState, useEffect } from 'react'
-import { userScores } from '../../UserScores';
+import React from 'react'
 
-export default function ScoreBoard({navigateBack}) {
-  const [players, setPlayers] = useState()
-
-  const URL = process.env.REACT_APP_URL || "localhost:5000/players"
-
-  useEffect(() => {
-    const getPlayers = async () => {
-      const res = await fetch(URL)
-      const data = await res.json()
-      setPlayers(data)
-    }
-    getPlayers()
-  }, [URL])
+export default function ScoreBoard({navigateBack, players}) {
 
   if (!players) return "Loading highscores..."
   
@@ -23,15 +10,17 @@ export default function ScoreBoard({navigateBack}) {
       <button className="btn-category" onClick={navigateBack}>Back</button>
       <h1 className="test">High Scores: </h1>
       <table align="center" >
+        <tbody>
         <tr>
-          <ol>
-            <td classname="scoreBoard">
-            {players.map((player, index) => {
-              return <li key={index}>{player.name.toUpperCase()} : {player.score} - {player.difficulty.toUpperCase()} </li>
-            })}
+            <td className="scoreBoard">
+              <ol>
+              {players.map((player, index) => {
+                return <li key={index}>{player.name.toUpperCase()} : {player.score} - {player.difficulty.toUpperCase()} </li>
+              })}
+              </ol>
             </td>
-          </ol>
         </tr>
+        </tbody>
       </table>
     </div>
   )
