+++ src/components/__tests__/ScoreBoard.test.js
@@ -1,26 +1,28 @@
-// import { fireEvent, render, screen } from "@testing-library/react";
-// import ScoreBoard from "../ScoreBoard/ScoreBoard";
+import {render, screen } from "@testing-library/react";
+import ScoreBoard from "../ScoreBoard/ScoreBoard";
 
-// const fakeScores = [
-//   {
-//     username: "test",
-//     score: 100
-//   },
-//   {
-//     username: "prueba",
-//     score: 80
-//   },
-//   {
-//     username: "appreuve",
-//     score: 60
-//   }
-// ];
+const fakeScores = [
+  {
+    name: "test",
+    score: 100,
+    difficulty: "easy"
+  },
+  {
+    name: "prueba",
+    score: 80,
+    difficulty: "medium"
+  },
+  {
+    name: "appreuve",
+    score: 60,
+    difficulty: "hard"
+  }
+];
 
 test("displays 3 items of userdata as an ordered list", () => {
-  // render(<ScoreBoard userScores={fakeScores} />);
-  // const listItems = screen.getAllByRole("listitem");
-  // expect(listItems[0]).toHaveTextContent("TEST");
-  // expect(listItems[1]).toHaveTextContent("PRUEBA");
-  // expect(listItems[2]).toHaveTextContent("APPREUVE");
-  expect(true).toBe(true);
+  render(<ScoreBoard players={fakeScores} />);
+  const listItems = screen.getAllByRole("listitem");
+  expect(listItems[0]).toHaveTextContent("TEST");
+  expect(listItems[1]).toHaveTextContent("PRUEBA");
+  expect(listItems[2]).toHaveTextContent("APPREUVE");
 });
